requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-extensions==3.2.3